Git ignore .idea folder, converted AngleLowPassFilter.java to a generic LowPassFilter, and added guardrails in the StateMachine
